requirements.txt
beautifulsoup4==4.9.0
openai==0.2.4
openai==0.27.4
python-dotenv==1.0.0
Requests==2.28.2
